Use 32-bit footprints for HAWAII to fit models with large feature maps

+++ common/data_structures.h
@@ -88,7 +88,7 @@ struct NodeFlags {
 static_assert(sizeof(struct NodeFlags) == NODE_FLAGS_SIZE + 2, "Unexpected size for NodeFlags");
 
 struct Footprint {
-    uint16_t value;
+    uint32_t value;
     uint8_t version;
     uint8_t dummy;
 };
